// generated/grpcServer/house_pb.d.ts
export class House extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setId(value?: google_protobuf_wrappers_pb.Int32Value): House;

    hasSquarefeet(): boolean;
    clearSquarefeet(): void;
    getSquarefeet(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setSquarefeet(value?: google_protobuf_wrappers_pb.Int32Value): House;

    hasNumberofbedrooms(): boolean;
    clearNumberofbedrooms(): void;
    getNumberofbedrooms(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setNumberofbedrooms(value?: google_protobuf_wrappers_pb.Int32Value): House;

    hasOnsale(): boolean;
    clearOnsale(): void;
    getOnsale(): google_protobuf_wrappers_pb.BoolValue | undefined;
    setOnsale(value?: google_protobuf_wrappers_pb.BoolValue): House;

    hasIsrental(): boolean;
    clearIsrental(): void;
    getIsrental(): google_protobuf_wrappers_pb.BoolValue | undefined;
    setIsrental(value?: google_protobuf_wrappers_pb.BoolValue): House;

    hasOwnerid(): boolean;
    clearOwnerid(): void;
    getOwnerid(): google_protobuf_wrappers_pb.Int32Value | undefined;
    setOwnerid(value?: google_protobuf_wrappers_pb.Int32Value): House;

    hasAddress(): boolean;
    clearAddress(): void;
    getAddress(): HouseAddress | undefined;
    setAddress(value?: HouseAddress): House;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): House.AsObject;
    static toObject(includeInstance: boolean, msg: House): House.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: House, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): House;
    static deserializeBinaryFromReader(message: House, reader: jspb.BinaryReader): House;
}

export namespace House {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.Int32Value.AsObject,
        squarefeet?: google_protobuf_wrappers_pb.Int32Value.AsObject,
        numberofbedrooms?: google_protobuf_wrappers_pb.Int32Value.AsObject,
        onsale?: google_protobuf_wrappers_pb.BoolValue.AsObject,
        isrental?: google_protobuf_wrappers_pb.BoolValue.AsObject,
        ownerid?: google_protobuf_wrappers_pb.Int32Value.AsObject,
        address?: HouseAddress.AsObject,
    }
}

export class HouseAddress extends jspb.Message { 

    hasStreetname(): boolean;
    clearStreetname(): void;
    getStreetname(): google_protobuf_wrappers_pb.StringValue | undefined;
    setStreetname(value?: google_protobuf_wrappers_pb.StringValue): HouseAddress;

    hasHousenumber(): boolean;
    clearHousenumber(): void;
    getHousenumber(): google_protobuf_wrappers_pb.StringValue | undefined;
    setHousenumber(value?: google_protobuf_wrappers_pb.StringValue): HouseAddress;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): HouseAddress.AsObject;
    static toObject(includeInstance: boolean, msg: HouseAddress): HouseAddress.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: HouseAddress, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): HouseAddress;
    static deserializeBinaryFromReader(message: HouseAddress, reader: jspb.BinaryReader): HouseAddress;
}

export namespace HouseAddress {
    export type AsObject = {
        streetname?: google_protobuf_wrappers_pb.StringValue.AsObject,
        housenumber?: google_protobuf_wrappers_pb.StringValue.AsObject,
    }
}